requirements.txt
streamlit
pandas
numpy
plotly
streamlit-reveal-slides